Improving FetchMetadata
Added:
- Use trakt if tmdb isn't find

Changed:
- Now connects to PMS with `ConnectPlexMediaServer()` method.
- While searching through tmdb results, if we have tmdb id and it is same as `entry["id"]` we correctly found matching content.

Currently for a show It wasn't giving the correct Metadata. Also id should be preferred to identify content than string title or year (those can vary).

Now, we get the guids of the media, if trakt is setup. We get the first database id in the session guids list and do a simple request to trakt, we later use the data returned for the button link. For the poster, trakt documentation says that they use tmdb, so we have to use it. Otherwise, we use the default search method.

**Why don't I have only used TMDB API?**
TMDB uses id depending of the media type, this makes things a bit harder but not much. True issue is that getting details for TV episode is a pure nightmare. Every database has a way of sorting episodes and seasons. Using id is made to bypass that and directly know which content it is. TMDB however, ask for the tmdb id, but also the season and episode number. Issue is that if you use another database than tmdb on plex you might get an index (episode) and parentIndex (season) that point to a non existing episode for tmdb and so no episode details. Here we at least use trakt to get link to the show.

I have tried to follow the file naming and type hinting convention as closely as possible, even if it doesn't conform to the PEP8 standard.